// packages/test/generatedCode/deviceAppManagement/mobileAppConfigurations/item/userStatuses/item/managedDeviceMobileAppConfigurationUserStatusItemRequestBuilderGetRequestConfiguration.ts
import {ManagedDeviceMobileAppConfigurationUserStatusItemRequestBuilderGetQueryParameters} from './managedDeviceMobileAppConfigurationUserStatusItemRequestBuilderGetQueryParameters';
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface ManagedDeviceMobileAppConfigurationUserStatusItemRequestBuilderGetRequestConfiguration {
    /** Request headers */
    headers?: Record<string, string>;
    /** Request options */
    options?: RequestOption[];
    /** Request query parameters */
    queryParameters?: ManagedDeviceMobileAppConfigurationUserStatusItemRequestBuilderGetQueryParameters;
}